requirements.txt
streamlit
google-cloud-storage
pandas
plotly